requirements.txt
streamlit
openai
langchain
langchain-community
langchainhub
langchain-openai
chromadb
python-dotenv
